Move utils from const_eval to mir_transform
