+++ compiler/rustc_borrowck/src/diagnostics/conflict_errors.rs
@@ -1846,7 +1846,7 @@ impl<'cx, 'tcx> MirBorrowckCtxt<'cx, 'tcx> {
             Some((method_did, method_substs)),
         ) = (
             &self.body[loan.reserve_location.block].terminator,
-            rustc_const_eval::util::find_self_call(
+            rustc_mir_transform::util::find_self_call(
                 tcx,
                 self.body,
                 loan.assigned_place.local,
